Avoid resetting specs after every gem install
This is done so that the spec just installed is available for require,
but we can add just that instead of resetting everything.

To fix this I restore a method that was deprecated and removed. I
checked git history and the reason was simplicity but by adding this we
can avoid some annoying behaviour like unnecessary warnings when
resetting specs.